requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
pytest==8.3.3
Werkzeug<3.0
